+++ kong/api/routes/debug.lua
@@ -5,10 +5,14 @@
 -- at https://konghq.com/enterprisesoftwarelicense/.
 -- [ END OF LICENSE 0867164ffc95e54f04670b5169c09574bdbd9bba ]
 
-local get_sys_filter_level            = require("ngx.errlog").get_sys_filter_level
+local cjson                           = require("cjson.safe")
 local set_log_level                   = require("resty.kong.log").set_log_level
+local get_log_level                   = require("resty.kong.log").get_log_level
+local constants                       = require("kong.constants")
 
-local LOG_LEVELS                      = require("kong.constants").LOG_LEVELS
+local LOG_LEVELS                      = constants.LOG_LEVELS
+local DYN_LOG_LEVEL_KEY               = constants.DYN_LOG_LEVEL_KEY
+local DYN_LOG_LEVEL_TIMEOUT_AT_KEY    = constants.DYN_LOG_LEVEL_TIMEOUT_AT_KEY
 
 local ngx                             = ngx
 local kong                            = kong
@@ -26,6 +30,7 @@ local ngx_worker_pid                  = ngx.worker.pid
 
 local NODE_LEVEL_BROADCAST            = false
 local CLUSTER_LEVEL_BROADCAST         = true
+local DEFAULT_LOG_LEVEL_TIMEOUT       = 60 -- 60s
 
 local MIN_PROFILING_TIMEOUT           = 10      -- seconds
 local MAX_PROFILING_TIMEOUT           = 600     -- seconds
@@ -58,6 +63,7 @@ local ERR_MSG_INVALID_INTERVAL        = string.format(
 
 local ERR_MSG_INVALID_MODE            = "invalid mode (must be 'time' or 'instruction'): "
 
+
 local function handle_put_log_level(self, broadcast)
   if kong.configuration.database == "off" then
     local message = "cannot change log level when not using a database"
@@ -65,27 +71,37 @@ local function handle_put_log_level(self, broadcast)
   end
 
   local log_level = LOG_LEVELS[self.params.log_level]
+  local timeout = math.ceil(tonumber(self.params.timeout) or DEFAULT_LOG_LEVEL_TIMEOUT)
 
   if type(log_level) ~= "number" then
     return kong.response.exit(400, { message = "unknown log level: " .. self.params.log_level })
   end
 
-  local sys_filter_level = get_sys_filter_level()
+  if timeout < 0 then
+    return kong.response.exit(400, { message = "timeout must be greater than or equal to 0" })
+  end
 
-  if sys_filter_level == log_level then
+  local cur_log_level = get_log_level(LOG_LEVELS[kong.configuration.log_level])
+
+  if cur_log_level == log_level then
     local message = "log level is already " .. self.params.log_level
     return kong.response.exit(200, { message = message })
   end
 
-  local ok, err = pcall(set_log_level, log_level)
+  local ok, err = pcall(set_log_level, log_level, timeout)
 
   if not ok then
     local message = "failed setting log level: " .. err
     return kong.response.exit(500, { message = message })
   end
 
+  local data = {
+    log_level = log_level,
+    timeout = timeout,
+  }
+
   -- broadcast to all workers in a node
-  ok, err = kong.worker_events.post("debug", "log_level", log_level)
+  ok, err = kong.worker_events.post("debug", "log_level", data)
 
   if not ok then
     local message = "failed broadcasting to workers: " .. err
@@ -94,7 +110,7 @@ local function handle_put_log_level(self, broadcast)
 
   if broadcast then
     -- broadcast to all nodes in a cluster
-    ok, err = kong.cluster_events:broadcast("log_level", tostring(log_level))
+    ok, err = kong.cluster_events:broadcast("log_level", cjson.encode(data))
 
     if not ok then
       local message = "failed broadcasting to cluster: " .. err
@@ -103,10 +119,17 @@ local function handle_put_log_level(self, broadcast)
   end
 
   -- store in shm so that newly spawned workers can update their log levels
-  ok, err = ngx.shared.kong:set("kong:log_level", log_level)
+  ok, err = ngx.shared.kong:set(DYN_LOG_LEVEL_KEY, log_level, timeout)
 
   if not ok then
     local message = "failed storing log level in shm: " .. err
+    return kong.response.exit(500, { message = message })
+  end
+
+  ok, err = ngx.shared.kong:set(DYN_LOG_LEVEL_TIMEOUT_AT_KEY, ngx.time() + timeout, timeout)
+
+  if not ok then
+    local message = "failed storing the timeout of log level in shm: " .. err
     return kong.response.exit(500, { message = message })
   end
 
@@ -128,15 +151,14 @@ end
 local routes = {
   ["/debug/node/log-level"] = {
     GET = function(self)
-      local sys_filter_level = get_sys_filter_level()
-      local cur_level = LOG_LEVELS[sys_filter_level]
+      local cur_level = get_log_level(LOG_LEVELS[kong.configuration.log_level])
 
-      if type(cur_level) ~= "string" then
-        local message = "unknown log level: " .. tostring(sys_filter_level)
+      if type(LOG_LEVELS[cur_level]) ~= "string" then
+        local message = "unknown log level: " .. tostring(cur_level)
         return kong.response.exit(500, { message = message })
       end
 
-      return kong.response.exit(200, { message = "log level: " .. cur_level })
+      return kong.response.exit(200, { message = "log level: " .. LOG_LEVELS[cur_level] })
     end,
   },
   ["/debug/node/log-level/:log_level"] = {
@@ -352,6 +374,7 @@ local routes = {
     end,
   },
 }
+
 
 local cluster_name