+++ kong/runloop/handler.lua
@@ -21,6 +21,7 @@ local lrucache     = require "resty.lrucache"
 local marshall     = require "kong.cache.marshall"
 local ktls         = require "resty.kong.tls"
 local workspaces   = require "kong.workspaces"
+local cjson        = require "cjson"
 
 local PluginsIterator = require "kong.runloop.plugins_iterator"
 local instrumentation = require "kong.tracing.instrumentation"
@@ -57,7 +58,7 @@ local escape            = require("kong.tools.uri").escape
 local is_http_module   = subsystem == "http"
 local is_stream_module = subsystem == "stream"
 
-
+local LOG_LEVELS                  = require("kong.constants").LOG_LEVELS
 local DEFAULT_MATCH_LRUCACHE_SIZE = Router.DEFAULT_MATCH_LRUCACHE_SIZE
 
 
@@ -86,7 +87,6 @@ local COMMA = byte(",")
 local SPACE = byte(" ")
 local QUESTION_MARK = byte("?")
 local ARRAY_MT = require("cjson.safe").array_mt
-local get_sys_filter_level = require("ngx.errlog").get_sys_filter_level
 
 local HOST_PORTS = {}
 
@@ -114,6 +114,7 @@ local STREAM_TLS_PASSTHROUGH_SOCK
 
 local set_authority
 local set_log_level
+local get_log_level
 local set_upstream_cert_and_key = ktls.set_upstream_cert_and_key
 local set_upstream_ssl_verify = ktls.set_upstream_ssl_verify
 local set_upstream_ssl_verify_depth = ktls.set_upstream_ssl_verify_depth
@@ -122,6 +123,7 @@ local set_upstream_ssl_trusted_store = ktls.set_upstream_ssl_trusted_store
 if is_http_module then
   set_authority = require("resty.kong.grpc").set_authority
   set_log_level = require("resty.kong.log").set_log_level
+  get_log_level = require("resty.kong.log").get_log_level
 end
 
 
@@ -939,10 +941,11 @@ return {
       if is_http_module then
         -- if worker has outdated log level (e.g. newly spawned), updated it
         timer_at(0, function()
-          local cur_log_level = get_sys_filter_level()
-          local shm_log_level = ngx.shared.kong:get("kong:log_level")
-          if cur_log_level and shm_log_level and cur_log_level ~= shm_log_level then
-            local ok, err = pcall(set_log_level, shm_log_level)
+          local cur_log_level = get_log_level(LOG_LEVELS[kong.configuration.log_level])
+          local shm_log_level = ngx.shared.kong:get(constants.DYN_LOG_LEVEL_KEY)
+          local timeout = (tonumber(ngx.shared.kong:get(constants.DYN_LOG_LEVEL_TIMEOUT_AT_KEY)) or 0) - ngx.time()
+          if shm_log_level and cur_log_level ~= shm_log_level and timeout > 0 then
+            local ok, err = pcall(set_log_level, shm_log_level, timeout)
             if not ok then
               local worker = ngx.worker.id()
               log(ERR, "worker" , worker, " failed setting log level: ", err)
@@ -959,7 +962,7 @@ return {
             return
           end
 
-          local ok, err = kong.worker_events.post("debug", "log_level", tonumber(data))
+          local ok, err = kong.worker_events.post("debug", "log_level", cjson.decode(data))
 
           if not ok then
             kong.log.err("failed broadcasting to workers: ", err)
@@ -975,14 +978,14 @@ return {
 
           log(NOTICE, "log level worker event received for worker ", worker)
 
-          local ok, err = pcall(set_log_level, data)
+          local ok, err = pcall(set_log_level, data.log_level, data.timeout)
 
           if not ok then
             log(ERR, "worker ", worker, " failed setting log level: ", err)
             return
           end
 
-          log(NOTICE, "log level changed to ", data, " for worker ", worker)
+          log(NOTICE, "log level changed to ", data.log_level, " for worker ", worker)
         end, "debug", "log_level")
       end