+++ scripts/explain_manifest/suites.py
@@ -36,7 +36,8 @@ def common_suites(expect, fips: bool = False):
         .functions \
         .contain("ngx_http_lua_kong_ffi_set_grpc_authority") \
         .contain("ngx_http_lua_ffi_balancer_enable_keepalive") \
-        .contain("ngx_http_lua_kong_ffi_set_log_level") \
+        .contain("ngx_http_lua_kong_ffi_set_dynamic_log_level") \
+        .contain("ngx_http_lua_kong_ffi_get_dynamic_log_level") \
         .contain("ngx_http_lua_kong_ffi_get_static_tag") \
         .contain("ngx_stream_lua_kong_ffi_get_static_tag") \
         .contain("ngx_http_lua_kong_ffi_get_full_client_certificate_chain") \
@@ -102,7 +103,7 @@ def ee_suites(expect, fips: bool = False):
         # no version requirements info for BoringSSL
 
 
-def libc_libcpp_suites(expect, max_libc: str, max_libcpp: str):
+def libc_libcpp_suites(expect, max_libc: str, max_libcxx: str, max_cxxabi: str):
     if max_libc:
         expect("**/*.so", "libc version is less than %s" % max_libc) \
             .version_requirement.key("libc.so.6").is_not().greater_than("GLIBC_%s" % max_libc) \
@@ -110,9 +111,13 @@ def libc_libcpp_suites(expect, max_libc: str, max_libcpp: str):
             .version_requirement.key("libpthread.so.0").is_not().greater_than("GLIBC_%s" % max_libc) \
             .version_requirement.key("librt.so.1").is_not().greater_than("GLIBC_%s" % max_libc) \
 
-    if max_libcpp:
-        expect("**/*.so", "libc version is less than %s" % max_libcpp) \
-            .version_requirement.key("libstdc++.so.6").is_not().greater_than("GLIBCXX_%s" % max_libcpp)
+    if max_libcxx:
+        expect("**/*.so", "glibcxx version is less than %s" % max_libcxx) \
+            .version_requirement.key("libstdc++.so.6").is_not().greater_than("GLIBCXX_%s" % max_libcxx)
+
+    if max_cxxabi:
+        expect("**/*.so", "cxxabi version is less than %s" % max_cxxabi) \
+            .version_requirement.key("libstdc++.so.6").is_not().greater_than("CXXABI_%s" % max_libcxx)
 
 
 def arm64_suites(expect):